--- a/docs/src/lib/internals.md
+++ b/docs/src/lib/internals.md
@@ -15,4 +15,5 @@ gennames
 getmaxwidths
 ourshow
 ourstrwidth
+tforeach
 ```