--- a/src/dataframerow/utils.jl
+++ b/src/dataframerow/utils.jl
@@ -25,13 +25,11 @@ function hashrows_col!(h::Vector{UInt},
                        v::AbstractVector{T},
                        rp::Nothing,
                        firstcol::Bool) where T
-    tforeach(eachindex(h), basesize=1_000_000) do i
-        @inbounds begin
-            el = v[i]
-            h[i] = hash(el, h[i])
-            if length(n) > 0
-                n[i] |= ismissing(el)
-            end
+    @inbounds for i in eachindex(h)
+        el = v[i]
+        h[i] = hash(el, h[i])
+        if length(n) > 0
+            n[i] |= ismissing(el)
         end
     end
     h
@@ -49,19 +47,19 @@ function hashrows_col!(h::Vector{UInt},
     # of the vector avoid using this path. 50% is roughly based on benchmarks
     if firstcol && Int64(2) * length(rp) < length(v)
         hashes = Vector{UInt}(undef, length(rp))
-        tforeach(eachindex(h), basesize=1_000_000) do i
-            @inbounds hashes[i] = hash(rp[i])
+        @inbounds for (i, v) in zip(eachindex(hashes), rp)
+            hashes[i] = hash(v)
         end
 
         fi = firstindex(rp)
         # here we rely on the fact that `DataAPI.refpool` has a continuous
         # block of indices
-        tforeach(enumerate(DataAPI.refarray(v)), basesize=1_000_000) do i, ref
-            @inbounds h[i] = hashes[ref+1-fi]
+        @inbounds for (i, ref) in enumerate(DataAPI.refarray(v))
+            h[i] = hashes[ref+1-fi]
         end
     else
-        tforeach(enumerate(v), basesize=1_000_000) do i
-            @inbounds h[i] = hash(x, h[i])
+        @inbounds for (i, x) in enumerate(v)
+            h[i] = hash(x, h[i])
         end
     end
     # Doing this step separately is faster, as it would disable SIMD above