@nalimilan 2021-03-14
>  I assume you have mostly copy-pasted the implementation. Right?

`tforeach` is a very simplified version inspired by `ThreadsX.foreach`, which was itself inspired by ThreadTools. But it's so simplified that it doesn't have a lot in common:
https://github.com/tkf/ThreadsX.jl/blob/6a69562f0fb5663e5661881698d9f11f7d45dc06/src/foreach.jl#L8